Added a Streaks Page
Cleaned up some of the logic in my battlegroups page as well as made it more mobile friendly and added a page that is taking our previous kill streak data from a csv and applying newly entered data to it. Also the KD entry page is currently fully functional and I am able to use it to track stats for Cav.