Added a lot of features
High streaks are now editable, so that you can store all time high streaks. Added a way to add a new season. Stats reset with new season and kill streaks continue through seasons. TDR is now completely live, so as deaths and kills are recorded the TDR will update with every 10 kills on that node automatically.